Add code to parse Day 7 tree structures
I didn't check the sample input for Day 7 before beginning to program,
and thought that the data was going to be pre-parsed as it is in the
example, rather than being constructed from terminal logs (which, in
retrospect, is probably actually a little bit easier than parsing the
tree structure). I wrote too much code to NOT save it somewhere.